Fixes: Made corrctions that ensued from the code review
