+++ XYZReader/src/main/java/com/example/xyzreader/adapter/RecyclerViewAdapter.java
@@ -1,0 +1,118 @@
+package com.example.xyzreader.adapter;
+
+import android.content.Context;
+import android.content.Intent;
+import android.database.Cursor;
+import android.support.v7.widget.RecyclerView;
+import android.text.Html;
+import android.text.format.DateUtils;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import com.example.xyzreader.R;
+import com.example.xyzreader.data.ArticleLoader;
+import com.example.xyzreader.data.ItemsContract;
+import com.example.xyzreader.ui.DynamicHeightNetworkImageView;
+import com.example.xyzreader.ui.ImageLoaderHelper;
+
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.GregorianCalendar;
+
+public class RecyclerViewAdapter extends RecyclerView.Adapter<ViewHolder> {
+
+    private static final String TAG = RecyclerViewAdapter.class.toString();
+
+    // Use default locale format
+    private SimpleDateFormat outputFormat = new SimpleDateFormat();
+    // Most time functions can only handle 1902 - 2037
+    private GregorianCalendar START_OF_EPOCH = new GregorianCalendar(2, 1, 1);
+    private SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.sss");
+
+    private Cursor mCursor;
+    private Context mContext;
+
+    public RecyclerViewAdapter(Context context, Cursor cursor) {
+        mContext = context;
+        mCursor = cursor;
+    }
+
+    @Override
+    public long getItemId(int position) {
+        mCursor.moveToPosition(position);
+        return mCursor.getLong(ArticleLoader.Query._ID);
+    }
+
+    private Date parsePublishedDate() {
+        try {
+            String date = mCursor.getString(ArticleLoader.Query.PUBLISHED_DATE);
+            return dateFormat.parse(date);
+        } catch (ParseException ex) {
+            Log.e(TAG, ex.getMessage());
+            Log.i(TAG, "passing today's date");
+            return new Date();
+        }
+    }
+
+    @Override
+    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        View view = LayoutInflater.from(mContext).inflate(R.layout.list_item_article, parent, false);
+        final ViewHolder vh = new ViewHolder(view);
+        view.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                mContext.startActivity(new Intent(Intent.ACTION_VIEW,
+                        ItemsContract.Items.buildItemUri(getItemId(vh.getAdapterPosition()))));
+            }
+        });
+        return vh;
+    }
+
+    @Override
+    public void onBindViewHolder(ViewHolder holder, int position) {
+        mCursor.moveToPosition(position);
+        holder.titleView.setText(mCursor.getString(ArticleLoader.Query.TITLE));
+        Date publishedDate = parsePublishedDate();
+        if (!publishedDate.before(START_OF_EPOCH.getTime())) {
+
+            holder.subtitleView.setText(Html.fromHtml(
+                    DateUtils.getRelativeTimeSpanString(
+                            publishedDate.getTime(),
+                            System.currentTimeMillis(), DateUtils.HOUR_IN_MILLIS,
+                            DateUtils.FORMAT_ABBREV_ALL).toString()
+                            + "<br/>" + " by "
+                            + mCursor.getString(ArticleLoader.Query.AUTHOR)));
+        } else {
+            holder.subtitleView.setText(Html.fromHtml(
+                    outputFormat.format(publishedDate)
+                            + "<br/>" + " by "
+                            + mCursor.getString(ArticleLoader.Query.AUTHOR)));
+        }
+        holder.thumbnailView.setImageUrl(
+                mCursor.getString(ArticleLoader.Query.THUMB_URL),
+                ImageLoaderHelper.getInstance(mContext).getImageLoader());
+        holder.thumbnailView.setAspectRatio(mCursor.getFloat(ArticleLoader.Query.ASPECT_RATIO));
+    }
+
+    @Override
+    public int getItemCount() {
+        return mCursor.getCount();
+    }
+}
+
+class ViewHolder extends RecyclerView.ViewHolder {
+    public DynamicHeightNetworkImageView thumbnailView;
+    public TextView titleView;
+    public TextView subtitleView;
+
+    public ViewHolder(View view) {
+        super(view);
+        thumbnailView = (DynamicHeightNetworkImageView) view.findViewById(R.id.thumbnail);
+        titleView = (TextView) view.findViewById(R.id.article_title);
+        subtitleView = (TextView) view.findViewById(R.id.article_subtitle);
+    }
+}
