+++ XYZReader/src/main/res/layout/list_item_article.xml
@@ -31,7 +31,7 @@
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:background="#B3FFFFFF"
+            android:background="@color/colorWhite_70_opacity"
             android:orientation="vertical"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
